Correct score mode in random sampler weight
